chore(871): snapshot test package.json
